django-template/requirements.txt
Django==4.1.7
django-environ==0.10.0
django-settings-export==1.2.1
django-debug-toolbar==3.8.1
django-import-export==3.1.0
django-cleanup==7.0.0
django-ipware==5.0.0
django-active-link==0.1.8
django-storages==1.13.2
django-cropperjs==0.0.5
django-simpleui==2023.3.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ipython==8.11.0
mysqlclient==2.1.1
Pillow==9.4.0
inflection
Werkzeug==2.2.3
boto3
requests==2.28.2
nplusone
pytz==2022.7.1
lambda_multiprocessing==0.5
python-docx==1.1.2
matplotlib==3.9.2
pandas==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
